Product cart update
Noticed we had a page for product details so I tied it to the existing cart functionality. With this we can add items from the front end.

Issue found is after an order the cart doesnt clear so the items appear in the cart again. Will need to fix this