BugFix
Update to fix some bugs.
app.jsx was calling the productDetails page and giving it an itemId causing an invalid product to be used, changed to productId to fix that. Updated the product cards to properly add to the cart and go to artist details. Added toastify notification to adding items so you get a nice little confirmation message now