update
Fixes for carts not clearing after an order and registration page now shows errors when the password is too short or the email already exists in the database.

Also added a field to the products table for artist_details. Ive updated the db.js file to include the field. updated abums_with_ids to include the field so should exist when you freshly install but you can also run
npm run backfill:bio to add it to your current db