Wishlist backend
Backend of wishlists working.

updated albums_with_ids.cjs to include every bands correct genres, since the fields in the db.js the backfill for genres is redundant.

Added a route for /public/images. album art will is now saved there and a relative URL is saved into the DB.